added hw4

+++ controllers/contacts/deleteContact.js
@@ -3,10 +3,11 @@ const { sendSuccessRes } = require('../../helpers')
 const { Contact } = require('../../models')
 
 const deleteContact = async (req, res) => {
+  const userId = req.user.id
   const { contactId } = req.params
-  const result = await Contact.findByIdAndDelete(contactId)
+  const result = await Contact.findByIdAndDelete({ _id: contactId, owner: userId })
   if (!result) {
-    throw new NotFound(404, `Contact with id=${contactId} not found`)
+    throw new NotFound(`Contact with id=${contactId} not found`)
   }
   sendSuccessRes(res, { message: 'Success delete' })
 }
